Extract datetime-local input format/parse helpers to src/utils
Pure TZ converters between UTC ISO and Bratislava-local HTML
datetime-local format. No behavior change; drops the `my` prefix since
they no longer live with the wrapper component.
